+++ src/main/java/io/fabric8/maven/docker/config/BuildImageConfiguration.java
@@ -1,7 +1,10 @@
 package io.fabric8.maven.docker.config;
 
 import java.io.File;
+import java.io.IOException;
 import java.io.Serializable;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.*;
 
 import io.fabric8.maven.docker.util.*;
@@ -19,14 +22,22 @@ public class BuildImageConfiguration implements Serializable {
     public static final String DEFAULT_CLEANUP = "try";
 
     /**
+     * Directory is used as build context.
+     * If not specified, dockerfile's parent directory is used as build context.
+     */
+    @Parameter
+    private String contextDir;
+
+    /**
      * Directory holding an external Dockerfile which is used to build the
      * image. This Dockerfile will be enriched by the addition build configuration
      */
     @Parameter
+    @Deprecated
     private String dockerFileDir;
 
     /**
-     * Path to a dockerfile to use. Its parent directory is used as build context (i.e. as <code>dockerFileDir</code>).
+     * Path to a dockerfile to use.
      * Multiple different Dockerfiles can be specified that way. If set overwrites a possibly givem
      * <code>dockerFileDir</code>
      */
@@ -148,6 +159,14 @@ public class BuildImageConfiguration implements Serializable {
 
     public boolean isDockerFileMode() {
         return dockerFileFile != null;
+    }
+
+    public File getContextDir() {
+        return contextDir != null ? new File(contextDir) : getDockerFile().getParentFile();
+    }
+
+    public String getContextDirRaw() {
+        return contextDir;
     }
 
     public File getDockerFile() {
@@ -306,6 +325,10 @@ public class BuildImageConfiguration implements Serializable {
         return args;
     }
 
+    public File getAbsoluteContextDirPath(MojoParameters mojoParams) {
+        return EnvUtil.prepareAbsoluteSourceDirPath(mojoParams, getContextDir().getPath());
+    }
+
     public File getAbsoluteDockerFilePath(MojoParameters mojoParams) {
         return EnvUtil.prepareAbsoluteSourceDirPath(mojoParams, getDockerFile().getPath());
     }
@@ -327,6 +350,11 @@ public class BuildImageConfiguration implements Serializable {
             } else {
                 this.config = DeepCopy.copy(that);
             }
+        }
+
+        public Builder contextDir(String dir) {
+            config.contextDir = dir;
+            return this;
         }
 
         public Builder dockerFileDir(String dir) {
@@ -547,19 +575,39 @@ public class BuildImageConfiguration implements Serializable {
     }
 
     private File findDockerFileFile(Logger log) {
+        if(dockerFileDir != null && contextDir != null) {
+            log.warn("Both contextDir (%s) and deprecated dockerFileDir (%s) are configured. Using contextDir.", contextDir, dockerFileDir);
+        }
+
         if (dockerFile != null) {
+            if (EnvUtil.isWindows() && !EnvUtil.isValidWindowsFileName(dockerFile)) {
+                throw new IllegalArgumentException(String.format("Invalid Windows file name %s for <dockerFile>", dockerFile));
+            }
+
             File dFile = new File(dockerFile);
-            if (dockerFileDir == null) {
+            if (dockerFileDir == null && contextDir == null) {
                 return dFile;
             } else {
-                if (dFile.isAbsolute()) {
-                    throw new IllegalArgumentException("<dockerFile> can not be absolute path if <dockerFileDir> also set.");
+                if(contextDir != null) {
+                    if (dFile.isAbsolute()) {
+                        return dFile;
+                    }
+                    return new File(contextDir, dockerFile);
                 }
-                if (EnvUtil.isWindows() && !EnvUtil.isValidWindowsFileName(dockerFile)) {
-                    throw new IllegalArgumentException(String.format("Invalid Windows file name %s for <dockerFile>", dockerFile));
+
+                if (dockerFileDir != null) {
+                    if (dFile.isAbsolute()) {
+                        throw new IllegalArgumentException("<dockerFile> can not be absolute path if <dockerFileDir> also set.");
+                    }
+                    log.warn("dockerFileDir parameter is deprecated, please migrate to contextDir");
+                    return new File(dockerFileDir, dockerFile);
                 }
-                return new File(dockerFileDir, dockerFile);
             }
+        }
+
+
+        if (contextDir != null) {
+            return new File(contextDir, "Dockerfile");
         }
 
         if (dockerFileDir != null) {