+++ doc/changelog.md
@@ -1,6 +1,6 @@
 # ChangeLog
 
-* 0.28.1
+* **0.28-SNAPSHOT**
   - Reintroduce minimal API-VERSION parameter in order to support docker versions below apiVersion 1.25
   - docs: Correct default image naming
   - Proxy settings are being ignored ([#1148](https://github.com/fabric8io/docker-maven-plugin/issues/1148))
@@ -12,7 +12,13 @@
   - Support docker SHELL setting for runCmds (#1157)
   - Added 'autoRemove' option for running containers (#1179)
   - Added support for AWS EC2 instance roles when pushing to AWS ECR (#1186)
+  - Introduce `contextDir` configuration option which would be used to specify docker build context (#1189)
   - Add support for auto-pulling multiple base image for multi stage builds (#1057)
+  - Fix usage of credential helper that do not support 'version' command (#1159)
+
+Please note that `dockerFileDir` is now deprecated in favor of `contextDir` which also allows absolute paths to Dockerfile with
+`dockerFile` and it will be removed in 1.0.0. It's still supported in this release but users are suggested to migrate to 
+`contextDir` instead.
   
 * **0.28.0** (2018-12-13)
   - Update to JMockit 1.43